Resolved lingering /types/types renameing for importation after refactor, and added in BASIC test template on new test files so they stop auto failing for lack of test
