+++ quell-server/__tests__/cacheOperations/invalidateCache.test.ts
@@ -17,3 +17,11 @@ ID cache reset
 Error handling
 Middleware next() calls
  */
+
+describe('Test Suite for invalidateCache.ts', () => {
+    console.log ('Test Suite for invalidateCache.ts');
+
+    it ('This is a test', () => {
+        console.log ('This is a filler test');
+    })
+})
