+++ quell-server/__tests__/cacheOperations/normalizeCache.test.ts
@@ -1,3 +1,11 @@
 /** 
 normalizeCache.ts (currently empty)
  */
+
+describe('Test Suite for normalizeCache.ts', () => {
+    console.log ('Test Suite for normalizeCache.ts');
+
+    it ('This is a test', () => {
+        console.log ('This is a filler test');
+    })
+})
